requirements.txt
--extra-index-url https://download.pytorch.org/whl/rocm6.3
torch~=2.7.1
torchvision~=0.22.1
torchaudio~=2.7.1
kagglehub>=0.2.3
matplotlib>=3.7.1
seaborn>=0.12.2
scikit-learn>=1.3.0
tqdm>=4.66.1
sympy~=1.14.0
numpy~=2.3.2
fastapi
uvicorn
python-multipart
pillow